chore : refacto progress bar
