--- src/components/ProgressBar/core/progressbar.typings.tsx
@@ -1,4 +1,0 @@
-export interface IProps {
-  value: number
-  max: number
-}
