+++ src/components/ProgressBar/internal/progress-bar.module.scss
